[Feature] sc-27764 Extract datasources as multiple python files
[Feature] sc-27764 Fix flake8 issues
